refactor: move shell-level chrome from canvas/ to editor/ root
Move six shell-level components from canvas/ to editor/ root with kebab-case names:
- CanvasEditorShell.tsx → canvas-editor-shell.tsx
- ViewChangesModal.tsx → view-changes-modal.tsx
- DeployPill.tsx → deploy-pill.tsx
- AutosaveStatus.tsx → autosave-status.tsx
- PublicEndpointRow.tsx → public-endpoint-row.tsx
- DraftTabPlaceholder.tsx → draft-tab-placeholder.tsx

Move five corresponding test files to editor/tests/ and fix all import specifiers
(absolute @/ paths and relative component imports) to match new locations.
Canvas surface files (StackCanvasTab, CanvasEditor, nodes, drawers, etc.) remain
in canvas/ for later refactor task.
