+++ frontend/src/pages/stacks/components/editor/draft-tab-placeholder.tsx
