+++ frontend/src/pages/stacks/components/editor/public-endpoint-row.tsx
