+++ frontend/src/pages/stacks/components/editor/view-changes-modal.tsx
